refactor: replace YAML config with challenges.json as single source of truth
